Name the eight that mattered, not all eighty
Personal figures have always been entered a whole roster at a time. That works for a five-a-side, where ten people fit on one screen. It does not work for a battle royale lobby: twenty squads of four is eighty rows for one fixture, and nobody types them, so the leaderboard stays empty.

A profile can now say how many players are named per match. Pressing Save opens a dialog with that many rows, each a dropdown of everyone in the fixture grouped by squad, and only the named ones are recorded. Left blank, nothing changes and the roster behaviour stands, so a bracket profile is untouched.

The setting sits with the stats it governs rather than on the event, because it only means anything where player tracking is on. That is the same switch that decides whether a sport has personal figures at all, so badminton and racing never see it.

Which reader runs is decided by the profile and never by what the browser sent. A request carrying the old roster keys to a profile that names players is ignored rather than trusted, and the squad a named player belongs to is worked out from their registration so figures cannot be filed under a team they do not play for.

A correction removes anybody no longer named. updateOrCreate alone was right for a fixed roster, where a blank row means no figure; here it would have left a player on the leaderboard with figures nobody claims after they were taken off the list.

A stat may now declare itself measured rather than counted. There is no such thing as two and a half kills, but a rating of 260.65 truncated to 260 loses exactly the part that separates two players. Opt-in on the component, so no existing profile changes behaviour because somebody typed a decimal point.

And the public side gets the podium a broadcast would show: second left, first raised in the middle, third right, with the figures on the plinth. One component serves the team table and the player table, and on a phone the winner comes first because there is no left and right to arrange.

+++ app/Models/TournamentMatchPlayer.php
@@ -69,9 +69,13 @@ class TournamentMatchPlayer extends Model
         return (float) data_get($this->component_points, $key, 0);
     }
 
-    public function componentCount(string $key): int
+    /**
+     * Float rather than int, for the same reason the standings accessor is: a personal
+     * figure can be measured as well as counted.
+     */
+    public function componentCount(string $key): float
     {
-        return (int) data_get($this->component_counts, $key, 0);
+        return (float) data_get($this->component_counts, $key, 0);
     }
 
     /**
